keep venue id and return url in state
