+++ lib/booking_ui/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/App.tsx":"1","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/index.tsx":"2","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LoginButton.tsx":"3","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/Profile.tsx":"4","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/reportWebVitals.ts":"5","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LogoutButton.tsx":"6","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/graph.ts":"7"},{"size":677,"mtime":1611187957343,"results":"8","hashOfConfig":"9"},{"size":1186,"mtime":1611187529591,"results":"10","hashOfConfig":"9"},{"size":351,"mtime":1611178968623,"results":"11","hashOfConfig":"9"},{"size":455,"mtime":1611176219838,"results":"12","hashOfConfig":"9"},{"size":425,"mtime":1611173987164,"results":"13","hashOfConfig":"9"},{"size":289,"mtime":1611175912174,"results":"14","hashOfConfig":"9"},{"size":11072,"mtime":1611186022451,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xpp6r5",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"23"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/App.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/index.tsx",[],["32","33"],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LoginButton.tsx",[],["34","35"],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/Profile.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/reportWebVitals.ts",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LogoutButton.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/graph.ts",[],{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","replacedBy":"39"},{"ruleId":"36","replacedBy":"40"},{"ruleId":"38","replacedBy":"41"},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],["42"],["43"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/App.tsx":"1","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/index.tsx":"2","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LoginButton.tsx":"3","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/Profile.tsx":"4","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/reportWebVitals.ts":"5","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LogoutButton.tsx":"6","/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/graph.ts":"7"},{"size":1489,"mtime":1611192986055,"results":"8","hashOfConfig":"9"},{"size":675,"mtime":1611190230638,"results":"10","hashOfConfig":"9"},{"size":377,"mtime":1611191929422,"results":"11","hashOfConfig":"9"},{"size":455,"mtime":1611176219838,"results":"12","hashOfConfig":"9"},{"size":425,"mtime":1611173987164,"results":"13","hashOfConfig":"9"},{"size":289,"mtime":1611175912174,"results":"14","hashOfConfig":"9"},{"size":11072,"mtime":1611186022451,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xpp6r5",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/App.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/index.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LoginButton.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/Profile.tsx",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/reportWebVitals.ts",[],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/components/LogoutButton.tsx",[],["31","32"],"/Users/matthew.cobbing/go/src/github.com/cobbinma/booking-platform/lib/booking_ui/src/graph.ts",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
