+++ lib/booking_ui/src/components/LoginButton.tsx
@@ -1,15 +1,12 @@
 import React from "react";
 import { useAuth0 } from "@auth0/auth0-react";
+import { Params } from "../App";
 
-const LoginButton = () => {
+const LoginButton: React.FC<{ params: Params }> = ({ params }) => {
   const { loginWithRedirect } = useAuth0();
 
   return (
-    <button
-      onClick={() =>
-        loginWithRedirect({ appState: { targetUrl: window.location.pathname } })
-      }
-    >
+    <button onClick={() => loginWithRedirect({ appState: { params: params } })}>
       Log In
     </button>
   );
